--- src/ContaPoupanca.kt
@@ -1,13 +1,0 @@
-class ContaPoupanca(
-    titular: String,
-    numero: Int
-) : Conta(
-    titular = titular,
-    numero = numero
-) {
-    override fun saca(valor: Double) {
-        if(this.saldo >= valor){
-            this.saldo -= valor
-        }
-    }
-}
