+++ src/main.kt
@@ -1,3 +1,5 @@
+import teste.testaAutenticacao
+
 fun main() {
     testaAutenticacao()
 }
