+++ src/modelo/Autenticavel.kt
@@ -1,3 +1,5 @@
+package modelo
+
 interface Autenticavel {
 
     fun autentica(senha: Int): Boolean
