+++ src/modelo/CalculadoraBonificacao.kt
@@ -1,3 +1,5 @@
+package modelo
+
 class CalculadoraBonificacao {
 
     var total: Double = 0.0
