+++ src/modelo/Conta.kt
@@ -1,3 +1,5 @@
+package modelo
+
 abstract class Conta(
     var titular: String,
     val numero: Int
