+++ src/modelo/Funcionario.kt
@@ -1,3 +1,5 @@
+package modelo
+
 abstract class Funcionario(
     val nome: String,
     val cpf: String,
