+++ src/modelo/FuncionarioAdmin.kt
@@ -1,3 +1,5 @@
+package modelo
+
 abstract class FuncionarioAdmin(
     nome: String,
     cpf: String,
