+++ src/modelo/Gerente.kt
@@ -1,9 +1,10 @@
-class Diretor(
+package modelo
+
+class Gerente(
     nome: String,
     cpf: String,
     salario: Double,
-    senha: Int,
-    val plr: Double
+    senha: Int
 ) : FuncionarioAdmin(
     nome = nome,
     cpf = cpf,
@@ -13,7 +14,7 @@ class Diretor(
 
     override val bonificacao: Double
         get() {
-            return salario + plr
+            return salario
         }
 
 }
