+++ src/teste/testaAutenticacao.kt
@@ -1,3 +1,10 @@
+package teste
+
+import modelo.Cliente
+import modelo.Diretor
+import modelo.Gerente
+import modelo.SistemaInterno
+
 fun testaAutenticacao() {
     val gerente = Gerente(
         nome = "Alex",
