+++ src/teste/testaLacos.kt
@@ -1,3 +1,5 @@
+package teste
+
 fun testaLacos() {
     var i = 0
     while (i < 5) {
